Registration collects name
First step to linking names to accounts is actually collecting the names
during registration. This customizes the registration form to collect
the name. This was *way* more difficult than I had expected. It seems
that yesod's auth system is fairly complicated and hard to use, unless
you are able to directly use one of the plugins provided in Yesod.Auth.
I found it strange that there was no plugin that allowed you to provide
a list of fields to collect during registration, so I may end up making
one on my own.